+++ Assets/Scripts/Items/Item.cs
@@ -1,8 +1,4 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
 public interface Item
 {
-    
+    int Id { get; }
 }
