+++ Assets/Scripts/Managers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 60405da6b9b04e349a33006c44b8fb94
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
