+++ Assets/Scripts/Managers/IGameManager.cs
