+++ Assets/Scripts/Managers/ManagerStatus.cs
