+++ Assets/Scripts/PlayerControlling.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be306388fa7e29d4fb547539f6fc3c5d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
